services/ingestion/requirements.txt
fastapi>=0.111,<0.112
uvicorn[standard]>=0.30,<0.31
apscheduler>=3.10.4,<3.11
pydantic-settings>=2.3,<2.4
boto3>=1.34,<1.35
botocore>=1.34,<1.35
python-dateutil>=2.9,<3
minio>=7.2,<7.3

# Inlined from services/radar-prepare/requirements.txt (cannot reference outside build context)
arm-pyart>=1.13.0
numpy>=1.20.0
rasterio>=1.2.0
